requirements.txt
google-generativeai
python-dotenv
colorama
termcolor
streamlit
